+++ src/components/profile/UserProductRegister.jsx
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import styled from 'styled-components';
 import { SERVER_URL } from '../../constants';
+import Product from '../common/Product';
 
 const BookContainer = styled.div`
 	display: flex;
@@ -14,69 +15,6 @@ const BookContainer = styled.div`
 		overflow-x: scroll;
 		overflow-y: hidden;
 		${({ theme }) => theme.ScrollbarStyle()}
-	}
-`;
-
-const BookList = styled.div`
-	cursor: pointer;
-`;
-
-const BookBtn = styled.button`
-	position: relative;
-	width: 31.5rem;
-	height: 25.2rem;
-	border: 0;
-	border-radius: 1rem;
-	background-size: cover;
-	background-position: center;
-	position: relative;
-	z-index: -1;
-	@media ${({ theme }) => theme.size.mobile} {
-		width: 15rem;
-		height: 12rem;
-	}
-	&::after {
-		content: '';
-		width: 100%;
-		height: 100%;
-		position: absolute;
-		border: 0;
-		border-radius: 1rem;
-		background-color: rgba(0, 0, 0, 0.6);
-		top: 0;
-		left: 0;
-		z-index: -1;
-	}
-`;
-
-const BookInfo = styled.div`
-	position: absolute;
-	bottom: 2.1rem;
-	left: 2.1rem;
-	text-align: left;
-	@media ${({ theme }) => theme.size.mobile} {
-		bottom: 1rem;
-		left: 1rem;
-	}
-`;
-
-const BookTitle = styled.p`
-	font-weight: 400;
-	font-size: 1.6rem;
-	color: ${({ theme }) => theme.grayColor5};
-	@media ${({ theme }) => theme.size.mobile} {
-		font-weight: 400;
-		font-size: 14px;
-	}
-`;
-
-const BookCost = styled.p`
-	font-weight: 700;
-	font-size: 1.4rem;
-	color: ${({ theme }) => theme.mainColor};
-	@media ${({ theme }) => theme.size.mobile} {
-		font-weight: 400;
-		font-size: 14px;
 	}
 `;
 
@@ -112,25 +50,7 @@ function UserProductRegister() {
 		<>
 			<BookContainer>
 				{userBooks?.map((item) => (
-					<BookList key={item.id}>
-						<BookBtn
-							type="button"
-							style={{
-								backgroundImage: `url(${item.itemImage})`,
-							}}
-						>
-							<BookInfo>
-								<BookTitle>{item.itemName}</BookTitle>
-								<BookCost>
-									{`${item.price}`.replace(
-										/\B(?<!\.\d*)(?=(\d{3})+(?!\d))/g,
-										',',
-									)}{' '}
-									원
-								</BookCost>
-							</BookInfo>
-						</BookBtn>
-					</BookList>
+					<Product key={item.id} item={item} />
 				))}
 			</BookContainer>
 		</>
